fs/squashfs: do not use CMD_RET_* defines in the filesystem code
Using CMD_RET_* defines in fs/squashfs/ breaks the build as they are
not defined. These defines are really meant to be used as return
values of commands, not internally in filesystem code. Instead,
appropriate errno codes should be used.

This patch changes fs/squashfs to use proper errno code. In most
places, it is straightforward. There are a few places where it is not:

 - In sqfs_read(), the value of datablk_count was checked *after* it
   was used to do a memcpy(). So the memcpy() could have used a
   negative size. The check was moved prior to the memcpy(), right
   after calling sqfs_get_regfile_info() and sqfs_get_lregfile_info().

 - The sqfs_size() function is modified to return "ret" at the end, so
   that the error code is properly propagated.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>