fs/squashfs: simplify superblock handling
Since we have the guarantee that the ->probe() function is called
before any operation, and that ->close() is called after any
operation, it means we can safely used that to store next to the
global cur_dev/cur_part_info variables the contents of the super block
that is read during ->probe().

This avoids reading the superblock a zillion times, and does not cause
any lifecycle issues as the superblock contents are read on ->probe()
and freed on ->close().

To make things a bit nicer, we enclose the complete global context
into a "struct squashfs_ctxt" structure, so that it is clear what
state gets set up during ->probe(), and cleaned up during ->close().

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>